EMCPDF1842: Change Long into BigInteger.

+++ dashboards-web/src/main/java/oracle/sysman/emaas/platform/dashboards/ws/rest/zdt/tablerows/DashboardFavoriteRowEntity.java
@@ -12,6 +12,8 @@ package oracle.sysman.emaas.platform.dashboards.ws.rest.zdt.tablerows;
 
 import org.codehaus.jackson.annotate.JsonProperty;
 
+import java.math.BigInteger;
+
 /**
  * @author guochen
  */
@@ -21,7 +23,7 @@ public class DashboardFavoriteRowEntity implements RowEntity
 	private String userName;
 
 	@JsonProperty("DASHBOARD_ID")
-	private Long dashboardId;
+	private BigInteger dashboardId;
 
 	@JsonProperty("CREATION_DATE")
 	private String creationDate;
@@ -102,7 +104,7 @@ public class DashboardFavoriteRowEntity implements RowEntity
 	/**
 	 * @return the dashboardId
 	 */
-	public Long getDashboardId()
+	public BigInteger getDashboardId()
 	{
 		return dashboardId;
 	}
@@ -161,7 +163,7 @@ public class DashboardFavoriteRowEntity implements RowEntity
 	 * @param dashboardId
 	 *            the dashboardId to set
 	 */
-	public void setDashboardId(Long dashboardId)
+	public void setDashboardId(BigInteger dashboardId)
 	{
 		this.dashboardId = dashboardId;
 	}
