+++ dashboards-web/src/main/java/oracle/sysman/emaas/platform/dashboards/ws/rest/zdt/tablerows/DashboardLastAccessRowEntity.java
@@ -12,13 +12,15 @@ package oracle.sysman.emaas.platform.dashboards.ws.rest.zdt.tablerows;
 
 import org.codehaus.jackson.annotate.JsonProperty;
 
+import java.math.BigInteger;
+
 /**
  * @author guochen
  */
 public class DashboardLastAccessRowEntity implements RowEntity
 {
 	@JsonProperty("DASHBOARD_ID")
-	private Long dashboardId;
+	private BigInteger dashboardId;
 
 	@JsonProperty("ACCESSED_BY")
 	private String accessedBy;
@@ -121,7 +123,7 @@ public class DashboardLastAccessRowEntity implements RowEntity
 	/**
 	 * @return the dashboardId
 	 */
-	public Long getDashboardId()
+	public BigInteger getDashboardId()
 	{
 		return dashboardId;
 	}
@@ -185,7 +187,7 @@ public class DashboardLastAccessRowEntity implements RowEntity
 	 * @param dashboardId
 	 *            the dashboardId to set
 	 */
-	public void setDashboardId(Long dashboardId)
+	public void setDashboardId(BigInteger dashboardId)
 	{
 		this.dashboardId = dashboardId;
 	}
